feat(v6): canonical track schema and the FFI provider surface
Add the typed data the V6 domains read from MusicBee, ahead of the
handlers that use it.

TrackTags carries the base browse fields plus the raw extended tags
MusicBee ships as strings (Year, Duration, Rating, DateAdded), read per
path via a new LibraryTrackTags query. The V6 `track` schema derives
its typed year / duration_ms / rating from these; DateAdded is
converted to ISO-8601 UTC on the C# side, where the running culture is
known, so the core never has to guess at a locale-ambiguous display
date (#114). It is deliberately separate from `Track`, which V4
`browsetracks` serializes byte-identically and so cannot grow fields.

Two more queries land with it: NowPlayingLyricsSynced (synchronised
LRC preferred, then unsynchronised, then the paramless now-playing
lyrics - a track can have synced lyrics the paramless call misses) and
HasLastFmAccount, so the core can answer `unavailable` rather than a
generic failure when scrobbling has no account behind it.

Also add an FFI-boundary smoke test that drives a query through a fake
C# `query_data` rejecting anything that is not a msgpack map. That
guards the `to_vec_named` contract a serde round-trip cannot: `to_vec`
emits a positional array, which compiles fine and throws on the first
live query.

Closes #112, #114
