+++ packages/mbrc-core/src/ffi/callbacks.rs
@@ -155,3 +155,113 @@ impl SafeCallbacks {
         }
     }
 }
+
+#[cfg(test)]
+mod ffi_boundary_smoke {
+    //! Guards the C#<->Rust MessagePack contract a pure serde round-trip CANNOT:
+    //! query params must serialize as a NAMED MAP (`to_vec_named`), not a positional
+    //! array (`to_vec`) - C#'s `ContractlessStandardResolver` reads structs as
+    //! name-keyed maps and throws "Unexpected msgpack code 147 (fixarray)" on an
+    //! array. A fake C# `query_data` here rejects anything that is not a map, so a
+    //! `to_vec` regression fails this test instead of only failing live.
+
+    use super::*;
+    use crate::ffi::types::{MbrcCallbacks, QueryType};
+    use serde::{Deserialize, Serialize};
+
+    #[derive(Serialize)]
+    struct Params {
+        offset: i32,
+        limit: i32,
+    }
+
+    #[derive(Serialize, Deserialize, Debug, PartialEq)]
+    struct Reply {
+        ok: bool,
+        n: i32,
+    }
+
+    /// True when the msgpack payload's first byte is a map marker (fixmap / map16 /
+    /// map32) - i.e. it was serialized with `to_vec_named`, as C# requires.
+    fn is_msgpack_map(bytes: &[u8]) -> bool {
+        matches!(bytes.first(), Some(b) if (0x80..=0x8f).contains(b) || *b == 0xde || *b == 0xdf)
+    }
+
+    /// Fake C# `query_data`: reject non-map params (status 1, as C# would throw),
+    /// else return a canned `Reply`. The result buffer intentionally leaks - a thin
+    /// `*mut u8` carries no length to reclaim (mirroring C#'s opaque allocator); it
+    /// is a few bytes in a test.
+    extern "C" fn fake_query(
+        _qt: i32,
+        params: *const u8,
+        params_len: u32,
+        out_buf: *mut *mut u8,
+        out_len: *mut u32,
+    ) -> i32 {
+        // SAFETY: the caller is the test below, which passes the pointer and
+        // length of a live msgpack buffer that outlives this call.
+        let slice = unsafe { std::slice::from_raw_parts(params, params_len as usize) };
+        if !is_msgpack_map(slice) {
+            return 1;
+        }
+        let mut reply = rmp_serde::to_vec_named(&Reply { ok: true, n: 7 }).unwrap();
+        reply.shrink_to_fit();
+        let len = reply.len() as u32;
+        let ptr = reply.as_mut_ptr();
+        std::mem::forget(reply);
+        // SAFETY: both out-params are non-null locals owned by the caller below,
+        // and the buffer they receive is leaked, so it outlives the write.
+        unsafe {
+            *out_buf = ptr;
+            *out_len = len;
+        }
+        0
+    }
+
+    extern "C" fn fake_free(_ptr: *mut u8) {
+        // No-op: see fake_query (the tiny test buffer leaks by design).
+    }
+
+    fn callbacks() -> SafeCallbacks {
+        SafeCallbacks::new(MbrcCallbacks {
+            query_data: Some(fake_query),
+            execute_command: None,
+            free_buffer: Some(fake_free),
+            on_event: None,
+        })
+    }
+
+    #[test]
+    fn query_params_cross_the_boundary_as_a_map() {
+        // If `query` ever used `to_vec`, the standin C# rejects the array and this
+        // is Err - catching the exact regression that is invisible to serde tests.
+        let reply: Reply = callbacks()
+            .query(
+                QueryType::PlayerState,
+                &Params {
+                    offset: 0,
+                    limit: 10,
+                },
+            )
+            .expect("map params accepted + reply round-tripped");
+        assert_eq!(reply, Reply { ok: true, n: 7 });
+    }
+
+    #[test]
+    fn to_vec_would_regress_to_an_array() {
+        // Document the landmine: to_vec_named is a map; to_vec is an array (the C#
+        // break). `query` must only ever emit the former.
+        let named = rmp_serde::to_vec_named(&Params {
+            offset: 0,
+            limit: 10,
+        })
+        .unwrap();
+        let positional = rmp_serde::to_vec(&Params {
+            offset: 0,
+            limit: 10,
+        })
+        .unwrap();
+        assert!(is_msgpack_map(&named), "to_vec_named must be a map");
+        assert!(!is_msgpack_map(&positional), "to_vec is an array");
+    }
+}
